testsite/requirements.txt
Django>=2.1.3,<2.2.0
psycopg2-binary==2.8.3
djangorestframework>=3.9.0,<3.10.0
gunicorn==19.9.0
Pillow==7.1.2
django-ckeditor

